backend/requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
websockets>=11.0

# Database - PostgreSQL with pgvector
psycopg2-binary>=2.9.9  # PostgreSQL adapter
pgvector>=0.2.0  # pgvector Python client
sqlalchemy>=2.0.0
alembic>=1.12.0

# Audio processing
faster-whisper>=0.9.0
librosa>=0.10.0
soundfile>=0.12.0

# Document processing
PyPDF2>=3.0.0
python-pptx>=0.6.20
python-docx>=0.8.11

# ML and embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0

# LLM integration
groq>=0.4.0
openai>=1.0.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Audio processing dependencies
nltk>=3.8.0
scipy>=1.10.0

# File handling
aiofiles>=23.0.0
pillow>=10.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
